+++ packages/event-tracker/jest.config.js
@@ -1,12 +1,7 @@
 const global = require('../../jest.config.global');
 
-const esModules = ['uuid'];
-
 module.exports = {
   ...global,
   displayName: 'event-tracker',
   coverageDirectory: '../../coverage/event-tracker',
-  transformIgnorePatterns: [
-    `/node_modules/(?!${esModules})`
-  ]
 };