+++ packages/avatar/CHANGELOG.md
@@ -2,6 +2,21 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.0](https://github.com/Availity/element/compare/@availity/mui-avatar@0.2.5...@availity/mui-avatar@0.3.0) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.2.5`
+
+### Features
+
+* add FileSelector ([7ad1f7b](https://github.com/Availity/element/commit/7ad1f7bb364bbeb2048d2ff4c9b0a2b1a1e33777))
+
+## [0.2.5](https://github.com/Availity/element/compare/@availity/mui-avatar@0.2.4...@availity/mui-avatar@0.2.5) (2024-10-25)
+
+### Dependency Updates
+
+* `design-tokens` updated to version `0.2.4`
 ## [0.2.4](https://github.com/Availity/element/compare/@availity/mui-avatar@0.2.3...@availity/mui-avatar@0.2.4) (2024-09-20)
 
 ### Dependency Updates