+++ packages/empty-state/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.13](https://github.com/Availity/element/compare/@availity/mui-empty-state@0.1.12...@availity/mui-empty-state@0.1.13) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-layout` updated to version `0.1.12`
+* `mui-link` updated to version `0.1.12`
 ## [0.1.12](https://github.com/Availity/element/compare/@availity/mui-empty-state@0.1.11...@availity/mui-empty-state@0.1.12) (2024-08-08)
 
 ### Dependency Updates