+++ packages/icon/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.11.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.10.1...@availity/mui-icon@0.11.0) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-layout` updated to version `0.10.1`
+
+### Features
+
+* add FileSelector ([7ad1f7b](https://github.com/Availity/element/commit/7ad1f7bb364bbeb2048d2ff4c9b0a2b1a1e33777))
+
 ## [0.10.1](https://github.com/Availity/element/compare/@availity/mui-icon@0.10.0...@availity/mui-icon@0.10.1) (2024-08-08)