+++ packages/modal/CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.8](https://github.com/Availity/element/compare/@availity/mui-modal@0.1.7...@availity/mui-modal@0.1.8) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-button` updated to version `0.1.7`
+* `mui-card` updated to version `0.1.7`
+* `mui-typography` updated to version `0.1.7`
 ## [0.1.7](https://github.com/Availity/element/compare/@availity/mui-modal@0.1.6...@availity/mui-modal@0.1.7) (2024-08-08)
 
 ### Dependency Updates