+++ packages/pagination/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.4](https://github.com/Availity/element/compare/@availity/mui-pagination@0.3.3...@availity/mui-pagination@0.3.4) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-icon` updated to version `0.3.3`
+* `mui-layout` updated to version `0.3.3`
+* `mui-list` updated to version `0.3.3`
+* `@availity/mock` updated to version `0.3.3`
+## [0.3.3](https://github.com/Availity/element/compare/@availity/mui-pagination@0.3.2...@availity/mui-pagination@0.3.3) (2024-10-25)
+
+### Dependency Updates
+
+* `mui-list` updated to version `0.3.2`
 ## [0.3.2](https://github.com/Availity/element/compare/@availity/mui-pagination@0.3.1...@availity/mui-pagination@0.3.2) (2024-10-15)
 
 ### Dependency Updates