+++ packages/textfield/CHANGELOG.md
@@ -2,6 +2,22 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.6.11](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.10...@availity/mui-textfield@0.6.11) (2024-10-31)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.6.10`
+* `mui-icon` updated to version `0.6.10`
+* `mui-button` updated to version `0.6.10`
+* `mui-chip` updated to version `0.6.10`
+* `mui-layout` updated to version `0.6.10`
+* `mui-menu` updated to version `0.6.10`
+## [0.6.10](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.9...@availity/mui-textfield@0.6.10) (2024-10-25)
+
+### Dependency Updates
+
+* `mui-form-utils` updated to version `0.6.9`
+* `mui-chip` updated to version `0.6.9`
 ## [0.6.9](https://github.com/Availity/element/compare/@availity/mui-textfield@0.6.8...@availity/mui-textfield@0.6.9) (2024-10-18)
 
 ### Dependency Updates